+++ R/dig_associations.R
@@ -1,7 +1,7 @@
 #' Search for association rules
 #'
 #' @description
-#' `r lifecycle::badge("deprecated")`
+#' `r lifecycle::badge("experimental")`
 #'
 #' Association rules identify conditions (*antecedents*) under which
 #' a specific feature (*consequent*) is present very often.
@@ -91,9 +91,8 @@
 #'                  min_support = 0.3,
 #'                  min_confidence = 0.8,
 #'                  measures = c("lift", "conviction"))
-#' @keywords internal
 #' @export
-dig_implications <- function(x,
+dig_associations <- function(x,
                              antecedent = everything(),
                              consequent = everything(),
                              disjoint = varnames(colnames(x)),
@@ -107,20 +106,145 @@ dig_implications <- function(x,
                              t_norm = "goguen",
                              threads = 1,
                              ...) {
-    lifecycle::deprecate_warn("1.4.0", "dig_implications()", with = "dig_associations()")
+    .must_be_double_scalar(min_coverage)
+    .must_be_in_range(min_coverage, c(0, 1))
 
-    dig_associations(x = x,
-                     antecedent = antecedent,
-                     consequent = consequent,
-                     disjoint = disjoint,
-                     min_length = min_length,
-                     max_length = max_length,
-                     min_coverage = min_coverage,
-                     min_support = min_support,
-                     min_confidence = min_confidence,
-                     contingency_table = contingency_table,
-                     measures = measures,
-                     t_norm = t_norm,
-                     threads = threads,
-                     ...)
+    .must_be_double_scalar(min_support)
+    .must_be_in_range(min_support, c(0, 1))
+
+    .must_be_double_scalar(min_confidence)
+    .must_be_in_range(min_confidence, c(0, 1))
+
+    .must_be_flag(contingency_table)
+    .must_be_enum(measures,
+                  values = c("lift", "conviction", "added_value"),
+                  null = TRUE,
+                  multi = TRUE)
+
+    min_coverage <- max(min_coverage, min_support)
+    n <- nrow(x)
+
+    antecedent <- enquo(antecedent)
+    consequent <- enquo(consequent)
+
+    f1 <- function(condition, support) {
+        res <- support
+        names(res) <- colnames(x)[condition]
+
+        res
+    }
+
+    conseq_supports <- dig(x = x,
+                           f = f1,
+                           condition = !!consequent,
+                           disjoint = disjoint,
+                           min_length = 1,
+                           max_length = 1,
+                           min_support = 0.0,
+                           threads = threads,
+                           error_context = list(arg_x = "x",
+                                                arg_condition = "consequent",
+                                                arg_disjoint = "disjoint",
+                                                arg_min_length = "min_length",
+                                                arg_max_length = "max_length",
+                                                arg_min_support = "min_support",
+                                                arg_threads = "threads",
+                                                call = current_env()))
+    conseq_supports <- unlist(conseq_supports)
+
+    basic_callback <- function(condition, sum, pp) {
+        conf <- pp / sum
+        sel <-!is.na(pp) & !is.na(conf) & conf >= min_confidence
+        selnames <- names(pp)[sel]
+        conf <- conf[sel]
+        supp <- pp[sel] / n
+        ante <- format_condition(names(condition))
+        cons <- unlist(lapply(names(conf), format_condition))
+
+        if (length(conf) <= 0) {
+            return(list(sel = logical(0), res = NULL))
+        }
+
+        list(sel = sel,
+             res =  data.frame(antecedent = ante,
+                               consequent = cons,
+                               support = supp,
+                               confidence = conf,
+                               coverage = sum / n,
+                               conseq_support = conseq_supports[selnames],
+                               count = pp[sel],
+                               antecedent_length = length(condition)))
+    }
+
+    f2 <- function(condition, sum, pp) {
+        basic_callback(condition, sum, pp)$res
+    }
+
+    f3 <- function(condition, sum, pp, pn, np, nn) {
+        bas <- basic_callback(condition, sum, pp)
+        sel <- bas$sel
+        res <- bas$res
+
+        if (length(res) <= 0) {
+            return(NULL)
+        }
+
+        res$pp <- pp[sel]
+        res$pn <- pn[sel]
+        res$np <- np[sel]
+        res$nn <- nn[sel]
+
+        res
+    }
+
+    contingency_needed_measures <- c("conviction")
+    contingency_needed <- length(intersect(measures, contingency_needed_measures)) > 0
+    f <- ifelse(contingency_table || contingency_needed, f3, f2)
+
+    res <- dig(x = x,
+               f = f,
+               condition = !!antecedent,
+               focus = !!consequent,
+               disjoint = disjoint,
+               min_length = min_length,
+               max_length = max_length,
+               min_support = min_coverage,
+               min_focus_support = min_support,
+               filter_empty_foci = TRUE,
+               t_norm = t_norm,
+               threads = threads,
+               error_context = list(arg_x = "x",
+                                    arg_condition = "antecedent",
+                                    arg_focus = "consequent",
+                                    arg_disjoint = "disjoint",
+                                    arg_min_length = "min_length",
+                                    arg_max_length = "max_length",
+                                    arg_min_support = "min_coverage",
+                                    arg_min_focus_support = "min_support",
+                                    arg_t_norm = "t_norm",
+                                    arg_threads = "threads",
+                                    call = current_env()),
+               ...)
+
+    res <- do.call(rbind, res)
+
+    if ("lift" %in% measures) {
+        res$lift <- res$support / (res$coverage * res$conseq_support)
+    }
+    if ("conviction" %in% measures) {
+        res$conviction <- res$coverage * (1 - res$conseq_support) / (res$pn / n)
+    }
+
+    if ("added_value" %in% measures) {
+        res$added_value <- res$confidence - res$conseq_support
+    }
+
+    if (!contingency_table) {
+        res$pp <- NULL
+        res$pn <- NULL
+        res$np <- NULL
+        res$nn <- NULL
+    }
+
+    as_tibble(res)
 }